refactor: move shared utility functions to utils. Fix some ESLint.
see pr description for ESLint fix
